refactor: Move old ts files
